% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavscat-package.R
\docType{package}
\name{wavscat-package}
\alias{wavscat}
\alias{wavscat-package}
\title{wavscat: Wavelet Scattering Transforms for Time Series and Audio}
\description{
Computes the wavelet scattering transform of one-dimensional signals: a cascade of Morlet wavelet convolutions, complex modulus and low-pass averaging that yields features stable to deformation and locally invariant to translation. Supports time scattering of order zero, one and two, and joint time-frequency scattering. The filter bank follows the construction of Anden and Mallat (2014) \doi{10.1109/TSP.2014.2326991} as implemented in the 'Kymatio' Python package. Everything is computed in the Fourier domain using base R, so the package has no compiled dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ABindoff/wavscat}
  \item Report bugs at \url{https://github.com/ABindoff/wavscat/issues}
}

}
\author{
\strong{Maintainer}: Aidan D. Bindoff \email{aidan.bindoff@utas.edu.au}

Authors:
\itemize{
  \item Aidan D. Bindoff \email{aidan.bindoff@utas.edu.au}
}

Other contributors:
\itemize{
  \item Kymatio developers (Filter bank and cascade design, BSD-3-Clause) [copyright holder]
}

}
\keyword{internal}
